fix libcurl on 32 bit systems
Curl expects size_t as the return type for the write function callback,
but freecwmp always returned a 64 bit int. This produced errors on big
endian 32 bit systems, as the upper 32 bit were always zero.

Contributed by T-Labs, Deutsche Telekom Innovation Laboratories

+++ src/http/http.c
@@ -166,7 +166,7 @@ done:
 }
 
 #ifdef HTTP_CURL
-static uint64_t
+static size_t
 http_get_response(void *buffer, size_t size, size_t rxed, char **msg_in)
 {
 	FC_DEVEL_DEBUG("enter");
